Move system simulation view into new view directory
