+++ src/tickit/views/system_simulation_view.py
